Move property inspector to own file and remove unnecessary code
