+++ Resources/Private/JavaScript/media-module/src/components/SideBarRight/Inspector/index.ts
@@ -2,6 +2,7 @@ import AssetInspector from './AssetInspector';
 import AssetCollectionInspector from './AssetCollectionInspector';
 import IptcMetadataInspector from './IptcMetadataInspector';
 import MetadataView from './MetadataView';
+import PropertyInspector from './PropertyInspector';
 import TagInspector from './TagInspector';
 import TagSelectBox from './TagSelectBox';
 import TagSelectBoxAsset from './TagSelectBoxAsset';
@@ -14,6 +15,7 @@ export {
     CollectionSelectBox,
     IptcMetadataInspector,
     MetadataView,
+    PropertyInspector,
     TagInspector,
     TagSelectBox,
     TagSelectBoxAsset,
